backend/requirements.txt
Django>=4.2
psycopg2-binary
selenium
djangorestframework
django-cors-headers
tabulate
